fix: 데이터 타입 수정
